refactor: 도메인 answer -> surveySubmission으로 변경
